fix(embed): parse csb package version number (#3010)
